third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <length> with iterationComposite
  assert_equals: expected "250px" but got "50px"
Harness: the test ran to completion.
